Split testDomains into rectangular and associative tests
