+++ frontend/test/resolution/CMakeLists.txt
@@ -33,7 +33,8 @@ comp_unit_test(testCustomInfer)
 comp_unit_test(testDeprecationUnstable)
 comp_unit_test(testDisambiguation)
 comp_unit_test(testDoWhile)
-comp_unit_test(testDomains)
+comp_unit_test(testDomainsRectangular)
+comp_unit_test(testDomainsAssociative)
 comp_unit_test(testEnums)
 comp_unit_test(testErrorContentLifetimes)
 comp_unit_test(testErrorGuard)
